requirements.txt
rich
PyYAML
python-dotenv
google-generativeai
whisperx
pyannote.audio
pydub
